+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 ## Unreleased
 ---
 Added
+- storage attribute on Container
+- Protocol.store() - manually change storage condition destiny of a Container
 
 Changed