+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 ## Unreleased
 ---
 Added
+
+Changed
+
+Removed
+
+Fixed
+
+## v2.2.2 - 2015-08-28
+Added
 - Storage attribute on Container
 - Protocol.store() - manually change storage condition destiny of a Container
 - Test for more complicated `transfer`ing with `one_source=True`
@@ -17,6 +26,7 @@ Removed
 
 Fixed
 - Error with `transfer`ing with `one_source=True`
+
 ## v2.2.1 - 2015-08-20
 ---
 Added